requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Django==5.0.6
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frontend==0.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
openai==1.37.1
packaging==24.1
psycopg2==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==71.1.0
sniffio==1.3.1
sqlparse==0.5.0
starlette==0.38.1
striprtf==0.0.26
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.3
